Unidad 1 y 2/PROYECTO_ED_APP/requirements.txt

reflex==0.8.26
bcrypt>=4.0.0
python-dotenv>=1.0.0